vim: textobj-indent: Fix a wrong column number
Fix a bug on the column number of the end of selection returned by
s:select().  The column number should be the length of the last line
+ 1, but the old code is wrong on two points;

(a) It uses col(), but it will never return any expected value.
(b) It calls col() with a line number, but such usage is invalid.